@@ -1,6 +1,7 @@
 package fr.alinedubois.mymeds.pharmacie.service;
 
-import javax.validation.constraints.NotNull;
+import jakarta.validation.constraints.NotNull;
+
 import java.time.LocalDate;
 
 public class AjoutMedicamentDTO {
